# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api/v1
    description: Versioned API base path
...
paths:
...
  /study-rooms:
    get:
      operationId: listStudyRooms
      tags: [studyRooms]
      summary: List study rooms
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
      responses:
        "200":
          description: Study room list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListStudyRoomsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createStudyRoom
      tags: [studyRooms]
      summary: Create study room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudyRoomRequest"
      responses:
        "201":
          description: Created study room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyRoomResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /study-rooms/{roomId}/goals:
    get:
      operationId: listStudyRoomGoals
      tags: [studyRooms]
      summary: List session goals for a study room
      parameters:
        - $ref: "#/components/parameters/RoomIdParam"
      responses:
        "200":
          description: Goals list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListStudyRoomGoalsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createStudyRoomGoal
      tags: [studyRooms]
      summary: Create a session goal
      parameters:
        - $ref: "#/components/parameters/RoomIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudyRoomGoalRequest"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyRoomGoalResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /study-rooms/{roomId}/goals/{goalId}:
    patch:
      operationId: updateStudyRoomGoal
      tags: [studyRooms]
      summary: Update a session goal
      parameters:
        - $ref: "#/components/parameters/RoomIdParam"
        - $ref: "#/components/parameters/GoalIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudyRoomGoalRequest"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyRoomGoalResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteStudyRoomGoal
      tags: [studyRooms]
      summary: Delete a session goal
      parameters:
        - $ref: "#/components/parameters/RoomIdParam"
        - $ref: "#/components/parameters/GoalIdParam"
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /study-rooms/{roomId}/timer:
    get:
      operationId: getStudyRoomTimer
      tags: [studyRooms]
      summary: Get shared timer state for a study room
      parameters:
        - $ref: "#/components/parameters/RoomIdParam"
      responses:
        "200":
          description: Timer state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyRoomTimerResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: patchStudyRoomTimer
      tags: [studyRooms]
      summary: Update shared timer state
      parameters:
        - $ref: "#/components/parameters/RoomIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchStudyRoomTimerRequest"
      responses:
        "200":
          description: Updated timer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyRoomTimerResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /study-rooms/{roomId}:
    get:
      operationId: getStudyRoomById
      tags: [studyRooms]
      summary: Get study room by id
      parameters:
        - $ref: "#/components/parameters/RoomIdParam"
      responses:
        "200":
          description: Study room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyRoomResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Study room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateStudyRoom
      tags: [studyRooms]
      summary: Update study room
      parameters:
        - $ref: "#/components/parameters/RoomIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudyRoomRequest"
      responses:
        "200":
          description: Updated study room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudyRoomResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Study room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteStudyRoom
      tags: [studyRooms]
      summary: Delete study room
      parameters:
        - $ref: "#/components/parameters/RoomIdParam"
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Study room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /ai/jobs:
    get:
      operationId: listAiJobs
      tags: [ai]
      summary: List AI jobs (placeholder)
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
      responses:
        "200":
          description: Jobs list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAiJobsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  parameters:
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
...
    RoomIdParam:
      name: roomId
      in: path
      required: true
      schema:
        type: string
...
    GoalIdParam:
      name: goalId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          const: false
        error:
          type: string
      required:
        - success
        - error
    DeleteData:
      type: object
      properties:
        id:
          type: string
        deleted:
          type: boolean
      required:
        - id
        - deleted
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/DeleteData"
      required:
        - success
        - data
...
    StudyRoom:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - description
        - isPublic
        - createdAt
        - updatedAt
    CreateStudyRoomRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        isPublic:
          type: boolean
      required:
        - name
    UpdateStudyRoomRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        isPublic:
          type: boolean
      minProperties: 1
    StudyRoomResponse:
      type: object
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/StudyRoom"
      required:
        - success
        - data
    ListStudyRoomsResponse:
      type: object
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/StudyRoom"
          required:
            - items
      required:
        - success
        - data
...
    StudyRoomGoal:
      type: object
      properties:
        id:
          type: string
        roomId:
          type: string
        text:
          type: string
        done:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - roomId
        - text
        - done
        - sortOrder
        - createdAt
        - updatedAt
    CreateStudyRoomGoalRequest:
      type: object
      properties:
        text:
          type: string
      required:
        - text
    UpdateStudyRoomGoalRequest:
      type: object
      properties:
        text:
          type: string
        done:
          type: boolean
      minProperties: 1
    StudyRoomGoalResponse:
      type: object
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/StudyRoomGoal"
      required:
        - success
        - data
    ListStudyRoomGoalsResponse:
      type: object
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/StudyRoomGoal"
          required:
            - items
      required:
        - success
        - data
    StudyRoomTimer:
      type: object
      properties:
        phase:
          type: string
          enum: [idle, focus, break]
        durationSec:
          type: integer
          nullable: true
        remainingSec:
          type: integer
        running:
          type: boolean
        leaderUserId:
          type: string
          nullable: true
        anchorEndsAt:
          type: string
          nullable: true
        updatedAt:
          type: string
      required:
        - phase
        - durationSec
        - remainingSec
        - running
        - leaderUserId
        - anchorEndsAt
        - updatedAt
    PatchStudyRoomTimerRequest:
      type: object
      properties:
        phase:
          type: string
          enum: [idle, focus, break]
        durationSec:
          type: integer
        remainingSec:
          type: integer
        running:
          type: boolean
      minProperties: 1
    StudyRoomTimerResponse:
      type: object
      properties:
        success:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/StudyRoomTimer"
      required:
        - success
        - data
...
    AiJob:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        status:
          type: string
      required:
        - id
        - type
        - status
    ListAiJobsResponse:
      type: object
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/AiJob"
          required:
            - items
      required:
        - success
        - data
...